Laptop/airflow/requirements.txt
apache-airflow==2.10.2
apache-airflow-providers-postgres
kafka-python
psycopg2-binary
pandas
